core: Add initial package support for Linux Qualcomm
This new linux-qcom package adds kernel support for the  Qualcomm
platforms. The kernel used for now is Johan's tree as it is currenly the
best kernel to run on a Lenovo Thinkpad X13s laptop.

Add a hack patch for allowing Xorg to work with external display.

Also add a few more kernel configs via a dedicated fragment.

Signed-off-by: Abel Vesa <abelvesa@kernel.org>